+++ services.php
@@ -135,6 +135,7 @@ use PublishPress\Future\Modules\Workflows\Domain\Steps\Triggers\Runners\OnCustom
 use PublishPress\Future\Modules\Workflows\Interfaces\WorkflowEngineInterface;
 use PublishPress\Future\Modules\Workflows\Migrations\V040500OnScheduledStepsSchema;
 use PublishPress\Future\Modules\Debug\Migrations\V04905DebugLogRequestId;
+use PublishPress\Future\Modules\Debug\Migrations\V04906DebugLogTimestampMilliseconds;
 
 return [
     ServicesAbstract::PLUGIN_VERSION => PUBLISHPRESS_FUTURE_VERSION,
@@ -687,6 +688,10 @@ return [
                     $container->get(ServicesAbstract::DB_TABLE_WORKFLOW_SCHEDULED_STEPS_SCHEMA)
                 ),
                 new V04905DebugLogRequestId(
+                    $container->get(ServicesAbstract::HOOKS),
+                    $container->get(ServicesAbstract::DB_TABLE_DEBUG_LOG_SCHEMA)
+                ),
+                new V04906DebugLogTimestampMilliseconds(
                     $container->get(ServicesAbstract::HOOKS),
                     $container->get(ServicesAbstract::DB_TABLE_DEBUG_LOG_SCHEMA)
                 ),